+++ src/server/game/Entities/Unit/Unit.h
@@ -2380,7 +2380,7 @@ class Unit : public WorldObject
         uint32 CalcArmorReducedDamage(Unit* victim, const uint32 damage, SpellInfo const* spellInfo, WeaponAttackType attackType=MAX_ATTACK);
         void CalcAbsorbResist(Unit* victim, SpellSchoolMask schoolMask, DamageEffectType damagetype, const uint32 damage, uint32 *absorb, uint32 *resist, SpellInfo const* spellInfo = NULL);
         void CalcHealAbsorb(Unit* victim, const SpellInfo* spellProto, uint32 &healAmount, uint32 &absorb);
-        uint32 CalcAbsorb(Unit* victim, SpellInfo const* spellProto, uint32 amount);
+        float CalcAbsorb(Unit* victim, SpellInfo const* spellProto, float amount);
 
         void  UpdateSpeed(UnitMoveType mtype, bool forced);
         float GetSpeed(UnitMoveType mtype) const;
